Chore: Rename logging.py file
